requirements.txt
streamlit
firebase-admin
sentence-transformers
anthropic
numpy
pandas
scikit-learn
streamlit-agraph
python-frontmatter
pyYAML
